feat: change data preprocess functions to one script file and added example;
