--- notebooks/ken/requirements.txt
@@ -1,57 +1,0 @@
-asttokens==3.0.0
-colorama==0.4.6
-comm==0.2.2
-contourpy==1.3.2
-cycler==0.12.1
-debugpy==1.8.14
-decorator==5.2.1
-executing==2.2.0
-filelock==3.13.1
-fonttools==4.57.0
-fsspec==2024.6.1
-ipykernel==6.29.5
-ipython==9.1.0
-ipython_pygments_lexers==1.1.1
-jedi==0.19.2
-Jinja2==3.1.4
-joblib==1.4.2
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-kiwisolver==1.4.8
-MarkupSafe==2.1.5
-matplotlib==3.10.1
-matplotlib-inline==0.1.7
-mpmath==1.3.0
-nest-asyncio==1.6.0
-networkx==3.3
-numpy==2.2.4
-packaging==24.2
-pandas==2.2.3
-parso==0.8.4
-pillow==11.0.0
-platformdirs==4.3.7
-prompt_toolkit==3.0.51
-psutil==7.0.0
-pure_eval==0.2.3
-Pygments==2.19.1
-pyparsing==3.2.3
-python-dateutil==2.9.0.post0
-pytz==2025.2
-pywin32==310
-pyzmq==26.4.0
-scikit-learn==1.6.1
-scipy==1.15.2
-seaborn==0.13.2
-six==1.17.0
-stack-data==0.6.3
-sympy==1.13.1
-threadpoolctl==3.6.0
-torch==2.6.0+cu118
-torchaudio==2.6.0+cu118
-torchvision==0.21.0+cu118
-tornado==6.4.2
-traitlets==5.14.3
-typing_extensions==4.13.2
-tzdata==2025.2
-wcwidth==0.2.13
-xgboost==3.0.0
